refactor: 당첨 번호와 관련된 로직만 수행하도록 역할 조정
- Rank 계산 후 반환하는 메서드 삭제
- 매칭 결과 계산 조건 간소화
